requirements.txt
Django>=3.1.5,<3.2.0
djangorestframework>=3.12.2,<3.13.0
psycopg2>=2.8.6,<2.9.0
Pillow>=8.1.0,<8.2.0
django-rest-swagger==2.2.0
django-cors-headers>=3.7.0,<3.8.0
django-reversion>=3.0.9,<3.1.0
django-reversion-compare>=0.14.0,<1.0.0
drf-yasg==1.20.0
drf-nested-routers==0.93.3
django-admin-list-filter-dropdown==1.0.3

flake8>=3.8.4,<3.9.0